+++ CHANGELOG.md
@@ -16,6 +16,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Changed
 
 -   Improved the PerfSidebarSettingsNodes component (#402)
+-   Removed dynamic average (#407)
 
 ## [0.13.0] - 2022-02-17